requirements.txt
python-dotenv>=1.0.0
openai>=1.0.0
requests>=2.28.0
curl_cffi>=0.6.0
folium>=0.15.0
py-staticmaps>=0.4.0
Pillow>=9.0.0,<10.0.0
reverse-geocode>=1.6.0
